+++ lib/helper/image_generation_helper.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:get/get.dart';
 import 'package:matrix_ai/controller/generation_controller.dart';
 import 'package:matrix_ai/controller/image_generation_controller.dart';
@@ -36,13 +34,13 @@ class ImageGenerationHelper {
     required Function(String, {bool showAds}) generateImage,
   }) async {
     if (SubscriptionController.find.isPro) {
-      await _handleProOrAndroidUser(text, generateImage: generateImage);
+      await _handleProUser(text, generateImage: generateImage);
     } else {
       _handleFreeUser(text, generateImage: generateImage);
     }
   }
 
-  static Future<void> _handleProOrAndroidUser(String text,
+  static Future<void> _handleProUser(String text,
       {required Function(String, {bool showAds}) generateImage}) async {
     if (GenerationController.find.proUserLimitExceeded) {
       bool hasShowedFreeLimitDialog = await ImageGenerationController.find.hasShowedFreeLimitDialog();
@@ -74,16 +72,16 @@ class ImageGenerationHelper {
         generateImage(text, showAds: false);
       } else {
         // for now we are showing ads only on iOS devices
-        if (Platform.isAndroid) {
-          generateImage(text);
-        } else {
-          showAdsDialog(
-            onWatchAdPressed: () {
-              pop();
-              generateImage(text);
-            },
-          );
-        }
+        // if (Platform.isAndroid) {
+        //   generateImage(text);
+        // } else {
+        showAdsDialog(
+          onWatchAdPressed: () {
+            pop();
+            generateImage(text);
+          },
+        );
+        // }
       }
     }
   }