+++ lib/view/screens/root.dart
@@ -2,12 +2,12 @@
 import 'dart:async';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:matrix_ai/controller/aws_controller.dart';
 import 'package:matrix_ai/controller/review_controller.dart';
 import 'package:matrix_ai/controller/settings_controller.dart';
+import 'package:matrix_ai/imports.dart';
 import 'package:matrix_ai/view/screens/dashboard/dashboard.dart';
 import '../../controller/ads_controller.dart';
 import '../../controller/generation_controller.dart';
@@ -18,6 +18,7 @@ import '../../controller/subscription_controller.dart';
 import '../base/no_internet_dialog.dart';
 import '../../helper/ad.dart';
 import 'splash/splash.dart';
+import 'package:easy_audience_network/easy_audience_network.dart' as meta;
 
 class Root extends StatefulWidget {
   const Root({super.key});
@@ -29,6 +30,7 @@ class Root extends StatefulWidget {
 class _RootState extends State<Root> with WidgetsBindingObserver {
   bool _ready = false;
   AppOpenAd? _appOpenAd;
+  meta.InterstitialAd? _appOpenFacebook;
   Timer? openAdTimeout;
 
   bool _disconnected = false;
@@ -73,11 +75,19 @@ class _RootState extends State<Root> with WidgetsBindingObserver {
       await SubscriptionController.find.initialize().catchError((_) {});
 
       // show add
-      await _loadAppOpenAd().then((value) async {
-        await _appOpenAd?.showIfNotPro();
-      }).catchError((e) {
-        FirebaseCrashlytics.instance.recordError(e, StackTrace.current);
-      });
+      if (GetPlatform.isIOS) {
+        await _loadAppOpenAd().then((value) async {
+          await _appOpenAd?.showIfNotPro();
+        }).catchError((e) {
+          FirebaseCrashlytics.instance.recordError(e, StackTrace.current);
+        });
+      } else {
+        await _loadAppOpenAdFacebook().then((value) async {
+          await _appOpenFacebook?.show();
+        }).catchError((e) {
+          FirebaseCrashlytics.instance.recordError(e, StackTrace.current);
+        });
+      }
       //
       _ready = true;
       if (mounted) setState(() {});
@@ -106,17 +116,25 @@ class _RootState extends State<Root> with WidgetsBindingObserver {
   void didChangeAppLifecycleState(AppLifecycleState state) {
     if (state == AppLifecycleState.resumed) {
       if (_lastShownTime.difference(DateTime.now()).inMinutes > 5) {
-        _appOpenAd?.showIfNotPro().catchError((e) {});
+        if (GetPlatform.isIOS) {
+          _appOpenAd?.showIfNotPro().catchError((e) {});
+        } else {
+          _appOpenFacebook?.show().catchError((e) {});
+        }
         _lastShownTime = DateTime.now();
       }
     } else if (state == AppLifecycleState.paused) {
-      _loadAppOpenAd();
+      if (GetPlatform.isIOS) {
+        _loadAppOpenAd();
+      } else {
+        _loadAppOpenAdFacebook();
+      }
     }
     super.didChangeAppLifecycleState(state);
   }
 
   Future<AppOpenAd?> _loadAppOpenAd() async {
-    if (!SettingsController.find.showAppOpen) return null;
+    if (!SettingsController.find.showAppOpen || SubscriptionController.find.isPro) return null;
     openAdTimeout?.cancel();
 
     return AdsController.find.showAppOpenAd().then((value) {
@@ -127,6 +145,27 @@ class _RootState extends State<Root> with WidgetsBindingObserver {
             _appOpenAd?.dispose();
             _appOpenAd = null;
             _loadAppOpenAd();
+          },
+        );
+      } else {
+        openAdTimeout = Timer(const Duration(minutes: 1), _loadAppOpenAd);
+      }
+      return value;
+    });
+  }
+
+  Future<meta.InterstitialAd?> _loadAppOpenAdFacebook() async {
+    if (!SettingsController.find.showAppOpen || SubscriptionController.find.isPro) return null;
+    openAdTimeout?.cancel();
+
+    return AdsController.find.showAppOpenAdFacebook().then((value) {
+      if (value != null) {
+        _appOpenFacebook = value;
+        _appOpenFacebook?.listener = meta.InterstitialAdListener(
+          onDismissed: () {
+            _appOpenFacebook?.destroy();
+            _appOpenFacebook = null;
+            _loadAppOpenAdFacebook();
           },
         );
       } else {